feat: Add undo-tree plugin to project
This commit adds the undo-tree plugin to the project. The plugin is configured to use the telescope.nvim extension for enhanced undo/redo functionality. The undo-tree plugin is set up with a vertical layout strategy and a preview height of 0.4. The plugin also includes the scope.nvim extension for additional functionality.